+++ src/main/java/kitchenpos/menu/domain/repository/MenuRepository.java
@@ -1,6 +1,7 @@
-package kitchenpos.menu.domain.entity;
+package kitchenpos.menu.domain.repository;
 
 import java.util.List;
+import kitchenpos.menu.domain.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
