+++ src/main/java/kitchenpos/menugroup/domain/repository/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menugroup.domain.repository;
 
+import kitchenpos.menugroup.domain.entity.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
