+++ src/main/java/kitchenpos/order/domain/repository/OrderRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.order.domain.entity;
+package kitchenpos.order.domain.repository;
 
+import kitchenpos.order.domain.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
