+++ src/main/java/kitchenpos/order/service/OrderService.java
@@ -3,14 +3,14 @@ package kitchenpos.order.service;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import kitchenpos.menu.domain.entity.MenuRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.exception.NotFoundMenuException;
 import kitchenpos.order.domain.entity.Order;
 import kitchenpos.order.domain.entity.OrderLineItem;
-import kitchenpos.order.domain.entity.OrderRepository;
+import kitchenpos.order.domain.repository.OrderRepository;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.table.domain.entity.OrderTable;
-import kitchenpos.table.domain.entity.OrderTableRepository;
+import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.order.domain.value.OrderLineItems;
 import kitchenpos.order.domain.value.Quantity;
 import kitchenpos.order.dto.OrderRequest;
