+++ src/main/java/kitchenpos/product/domain/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.product.domain.entity;
+package kitchenpos.product.domain.repository;
 
+import kitchenpos.product.domain.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
