+++ src/main/java/kitchenpos/table/domain/repository/OrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.table.domain.entity;
+package kitchenpos.table.domain.repository;
 
+import kitchenpos.table.domain.entity.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
