+++ src/main/java/kitchenpos/table/domain/repository/TableGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.table.domain.entity;
+package kitchenpos.table.domain.repository;
 
+import kitchenpos.table.domain.entity.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
