+++ src/main/java/kitchenpos/table/service/TableGroupService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import kitchenpos.table.domain.entity.OrderTable;
-import kitchenpos.table.domain.entity.OrderTableRepository;
+import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.table.domain.entity.TableGroup;
-import kitchenpos.table.domain.entity.TableGroupRepository;
+import kitchenpos.table.domain.repository.TableGroupRepository;
 import kitchenpos.table.domain.value.OrderTables;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.TableGroupRequest;
