+++ src/test/java/kitchenpos/menu/service/MenuServiceTest.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.menu.domain.entity.Menu;
 import kitchenpos.menu.domain.value.MenuProducts;
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.entity.MenuGroup;
 import kitchenpos.menu.domain.entity.MenuProduct;
 import kitchenpos.menu.domain.value.Price;
 import kitchenpos.menugroup.dto.MenuGroupRequest;
